Create Finding_the_Best_Hospital
Attempting to create a subset from a data frame that allows me to input a State and the specified outcome(heart failure, heart attack, or pneumonia) and output the name of the hospital with the lowest mortality rate.
This one has errors still and is incomplete. Please don't rely on it for answers. I barely got through swirl.